Moved extract_year_from_date to utils and updated requirements.txt
