+++ requirements.txt
@@ -1,3 +1,4 @@
 discid
 mutagen>=1.37
 PyQt5>=5.7.1
+dateutils==2.8.1
